+++ lib/shoulda/matchers/active_model/validator_with_captured_range_error.rb
@@ -1,0 +1,11 @@
+module Shoulda
+  module Matchers
+    module ActiveModel
+      module ValidatorWithCapturedRangeError
+        def messages_description
+          ' RangeError: ' + captured_range_error.message.inspect
+        end
+      end
+    end
+  end
+end