+++ lib/shoulda/matchers/rails_shim.rb
@@ -65,6 +65,28 @@ module Shoulda
         end
       end
 
+      def self.generate_validation_message(record, attribute, type, model_name, options)
+        if record && record.errors.respond_to?(:generate_message)
+          record.errors.generate_message(attribute.to_sym, type, options)
+        else
+          simply_generate_validation_message(attribute, type, model_name, options)
+        end
+      rescue RangeError
+        simply_generate_validation_message(attribute, type, model_name, options)
+      end
+
+      def self.simply_generate_validation_message(attribute, type, model_name, options)
+        default_translation_keys = [
+          :"activerecord.errors.models.#{model_name}.#{type}",
+          :"activerecord.errors.messages.#{type}",
+          :"errors.attributes.#{attribute}.#{type}",
+          :"errors.messages.#{type}"
+        ]
+        primary_translation_key = :"activerecord.errors.models.#{model_name}.attributes.#{attribute}.#{type}"
+        translate_options = { default: default_translation_keys }.merge(options)
+        I18n.translate(primary_translation_key, translate_options)
+      end
+
       def self.active_record_major_version
         ::ActiveRecord::VERSION::MAJOR
       end