+++ spec/support/unit/helpers/active_model_helpers.rb
@@ -7,8 +7,10 @@ module UnitTests
     def custom_validation(options = {}, &block)
       attribute_name = options.fetch(:attribute_name, :attr)
       attribute_type = options.fetch(:attribute_type, :integer)
+      column_options = options.fetch(:column_options, {})
+      attribute_options = { type: attribute_type, options: column_options }
 
-      define_model(:example, attribute_name => attribute_type) do
+      define_model(:example, attribute_name => attribute_options) do
         validate :custom_validation
 
         define_method(:custom_validation, &block)