+++ spec/support/unit/helpers/rails_versions.rb
@@ -6,19 +6,19 @@ module UnitTests
     end
 
     def rails_version
-      Gem::Version.new(Rails::VERSION::STRING)
+      Tests::Version.new(Rails::VERSION::STRING)
     end
 
     def rails_3_x?
-      Gem::Requirement.new('~> 3.0').satisfied_by?(rails_version)
+      rails_version =~ '~> 3.0'
     end
 
     def rails_4_x?
-      Gem::Requirement.new('~> 4.0').satisfied_by?(rails_version)
+      rails_version =~ '~> 4.0'
     end
 
     def rails_gte_4_1?
-      Gem::Requirement.new('>= 4.1').satisfied_by?(rails_version)
+      rails_version >= 4.1
     end
 
     def active_record_supports_enum?